Fix compilation warnings.

+++ src/refine.c
@@ -7,7 +7,11 @@
 *
 *	This file part of:	SExtractor
 *
-*	Copyright:		(C) 1993-2016 Emmanuel Bertin -- IAP/CNRS/UPMC
+*	Copyright:		(C) 1994,1997 ESO
+*	          		(C) 1995,1996 Leiden Observatory 
+*	          		(C) 1998-2021 IAP/CNRS/SorbonneU
+*	          		(C)	2021-2023 CFHT/CNRS
+*	          		(C) 2023-2025 CEA/AIM/UParisSaclay
 *
 *	License:		GNU General Public License
 *
@@ -22,7 +26,7 @@
 *	You should have received a copy of the GNU General Public License
 *	along with SExtractor. If not, see <http://www.gnu.org/licenses/>.
 *
-*	Last modified:		13/01/2016
+*	Last modified:		19/03/2025
 *
 *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 
@@ -327,7 +331,7 @@ int	gatherup(objliststruct *objlistin, objliststruct *objlistout)
         }			
       if (p[nobj-1] > 1.0e-31)
         {
-        drand = p[nobj-1]*rand()/RAND_MAX;
+        drand = p[nobj-1]*rand()/(float)RAND_MAX;
         for (i=1; i<nobj && p[i]<drand; i++);
         if (i==nobj)
           i=iclst;
